+++ Knapsack.py
@@ -3,6 +3,13 @@ import random
 import tkinter as tk
 from tkinter import *
 import threading
+import logging
+
+# Configure logging
+logging.basicConfig(
+    level=logging.DEBUG, format="%(asctime)s [%(levelname)s]: %(message)s"
+)
+logger = logging.getLogger(__name__)
 
 num_items = 100
 frac_target = 0.7
@@ -13,19 +20,14 @@ screen_padding = 25
 item_padding = 5
 stroke_width = 5
 
-num_generations = 1000
-pop_size = 50
-elitism_count = 2
-mutation_rate = 0.1
-
 sleep_time = 0.1
 
 
 def random_rgb_color():
-    red = random.randint(0x10, 0xff)
-    green = random.randint(0x10, 0xff)
-    blue = random.randint(0x10, 0xff)
-    hex_color = '#{:02x}{:02x}{:02x}'.format(red, green, blue)
+    red = random.randint(0x10, 0xFF)
+    green = random.randint(0x10, 0xFF)
+    blue = random.randint(0x10, 0xFF)
+    hex_color = "#{:02x}{:02x}{:02x}".format(red, green, blue)
     return hex_color
 
 
@@ -45,90 +47,98 @@ class Item:
         self.h = h
 
     def draw(self, canvas, active=False):
-        canvas.create_text(self.x+self.w+item_padding+stroke_width*2, self.y+self.h/2, text=f'{self.value}')
+        canvas.create_text(
+            self.x + self.w + item_padding + stroke_width * 2,
+            self.y + self.h / 2,
+            text=f"{self.value}",
+        )
         if active:
-            canvas.create_rectangle(self.x,
-                                    self.y,
-                                    self.x+self.w,
-                                    self.y+self.h,
-                                    fill=self.color,
-                                    outline=self.color,
-                                    width=stroke_width)
+            canvas.create_rectangle(
+                self.x,
+                self.y,
+                self.x + self.w,
+                self.y + self.h,
+                fill=self.color,
+                outline=self.color,
+                width=stroke_width,
+            )
         else:
-            canvas.create_rectangle(self.x,
-                                    self.y,
-                                    self.x+self.w,
-                                    self.y+self.h,
-                                    fill='',
-                                    outline=self.color,
-                                    width=stroke_width)
+            canvas.create_rectangle(
+                self.x,
+                self.y,
+                self.x + self.w,
+                self.y + self.h,
+                fill="",
+                outline=self.color,
+                width=stroke_width,
+            )
 
 
 class UI(tk.Tk):
     def __init__(self):
-        tk.Tk.__init__(self)
-        # Set the title of the window
+        super().__init__()
         self.title("Knapsack")
-        # Hide the minimize/maximize/close decorations at the top of the window frame
-        #   (effectively making it act like a full-screen application)
         self.option_add("*tearOff", FALSE)
-        # Get the screen width and height
         self.width, self.height = self.winfo_screenwidth(), self.winfo_screenheight()
-        # Set the window width and height to fill the screen
         self.geometry("%dx%d+0+0" % (self.width, self.height))
-        # Set the window content to fill the width * height area
         self.state("zoomed")
 
         self.canvas = Canvas(self)
         self.canvas.place(x=0, y=0, width=self.width, height=self.height)
 
         self.items_list = []
-
-        # We create a standard banner menu bar and attach it to the window
-        menu_bar = Menu(self)
-        self['menu'] = menu_bar
-
-        # We have to individually create the "File", "Edit", etc. cascade menus, and this is the first
-        menu_K = Menu(menu_bar)
-        # The underline=0 parameter doesn't actually do anything by itself,
-        #   but if you also create an "accelerator" so that users can use the standard alt+key shortcuts
-        #   for the menu, it will underline the appropriate key to indicate the shortcut
-        menu_bar.add_cascade(menu=menu_K, label='Knapsack', underline=0)
-
-        def generate():
-            self.generate_knapsack()
-            self.draw_items()
-        # The add_command function adds an item to a menu, as opposed to add_cascade which adds a sub-menu
-        # Note that we use command=generate without the () - we're telling it which function to call,
-        #   not actually calling the function as part of the add_command
-        menu_K.add_command(label="Generate", command=generate, underline=0)
-
         self.target = 0
 
+        menu_bar = Menu(self)
+        self["menu"] = menu_bar
+
+        menu_K = Menu(menu_bar)
+        menu_bar.add_cascade(menu=menu_K, label="Knapsack", underline=0)
+
+        def generate():
+            logger.debug("Generating knapsack items...")
+            self.generate_knapsack()
+            self.draw_items()
+            logger.debug("Knapsack generation completed.")
+
+        menu_K.add_command(label="Generate", command=generate, underline=0)
+
         def set_target():
+            if not self.items_list:
+                logger.warning("No items available to set a target.")
+                return
+            logger.debug("Setting target...")
             target_set = []
             for x in range(int(num_items * frac_target)):
-                item = self.items_list[random.randint(0, len(self.items_list)-1)]
+                item = self.items_list[random.randint(0, len(self.items_list) - 1)]
                 while item in target_set:
                     item = self.items_list[random.randint(0, len(self.items_list) - 1)]
                 target_set.append(item)
-            total = 0
-            for item in target_set:
-                total += item.value
+            total = sum(i.value for i in target_set)
             self.target = total
             self.draw_target()
+            logger.debug(f"Target set to {self.target}")
+
         menu_K.add_command(label="Get Target", command=set_target, underline=0)
 
         def start_thread():
+            if not self.items_list:
+                logger.warning(
+                    "No items to run the algorithm on. Please generate items first."
+                )
+                return
+            if self.target == 0:
+                logger.warning("Target not set. Please set a target first.")
+                return
+            logger.debug("Starting DP solver in a separate thread...")
             thread = threading.Thread(target=self.run, args=())
             thread.start()
-        menu_K.add_command(label="Run", command=start_thread, underline=0)
 
-        # We have to call self.mainloop() in our constructor (__init__) to start the UI loop and display the window
-        self.mainloop()
+        menu_K.add_command(label="Run", command=start_thread, underline=0)
 
     def get_rand_item(self):
         i1 = Item()
+        # Ensuring uniqueness by value (although not strictly required)
         for i2 in self.items_list:
             if i1.value == i2.value:
                 return None
@@ -139,23 +149,22 @@ class UI(tk.Tk):
         while item is None:
             item = self.get_rand_item()
         self.items_list.append(item)
+        logger.debug(f"Item added: Value={item.value}, Color={item.color}")
 
     def generate_knapsack(self):
+        self.items_list.clear()
+        logger.debug("Generating items...")
         for i in range(num_items):
             self.add_item()
 
-        item_max = 0
-        item_min = 9999
-        for item in self.items_list:
-            item_min = min(item_min, item.value)
-            item_max = max(item_max, item.value)
-
+        item_max = max(item.value for item in self.items_list)
         w = self.width - screen_padding
         h = self.height - screen_padding
         num_rows = math.ceil(num_items / 6)
         row_w = w / 8 - item_padding
         row_h = (h - 200) / num_rows
-        # print(f'{w}, {h}, {num_rows}, {row_w}, {row_h}')
+
+        logger.debug("Placing items on the canvas...")
         for x in range(0, 6):
             for y in range(0, num_rows):
                 if x * num_rows + y >= num_items:
@@ -163,174 +172,120 @@ class UI(tk.Tk):
                 item = self.items_list[x * num_rows + y]
                 item_w = row_w / 2
                 item_h = max(item.value / item_max * row_h, 1)
-                # print(f'{screen_padding+x*row_w+x*item_padding},'
-                #      f'{screen_padding+y*row_h+y*item_padding},'
-                #      f'{item_w},'
-                #      f'{item_h}')
-                item.place(screen_padding + x * row_w + x * item_padding,
-                           screen_padding + y * row_h + y * item_padding,
-                           item_w,
-                           item_h)
+                item.place(
+                    screen_padding + x * row_w + x * item_padding,
+                    screen_padding + y * row_h + y * item_padding,
+                    item_w,
+                    item_h,
+                )
 
     def clear_canvas(self):
         self.canvas.delete("all")
 
-    def draw_items(self):
-        for item in self.items_list:
-            item.draw(self.canvas)
+    def draw_items(self, active_items=None):
+        if active_items is None:
+            active_items = []
+        for i, item in enumerate(self.items_list):
+            item.draw(self.canvas, i in active_items)
 
     def draw_target(self):
         x = (self.width - screen_padding) / 8 * 7
         y = screen_padding
         w = (self.width - screen_padding) / 8 - screen_padding
         h = self.height / 2 - screen_padding
-        self.canvas.create_rectangle(x, y, x + w, y + h, fill='black')
-        self.canvas.create_text(x+w//2, y+h+screen_padding, text=f'{self.target}', font=('Arial', 18))
+        self.canvas.create_rectangle(x, y, x + w, y + h, fill="black")
+        self.canvas.create_text(
+            x + w // 2,
+            y + h + screen_padding,
+            text=f"{self.target}",
+            font=("Arial", 18),
+        )
 
     def draw_sum(self, item_sum, target):
         x = (self.width - screen_padding) / 8 * 6
         y = screen_padding
         w = (self.width - screen_padding) / 8 - screen_padding
         h = self.height / 2 - screen_padding
-        # print(f'{item_sum} / {target} * {h} = {item_sum/target} * {h} = {item_sum/target*h}')
-        h *= (item_sum / target)
-        self.canvas.create_rectangle(x, y, x + w, y + h, fill='black')
-        self.canvas.create_text(x+w//2, y+h+screen_padding, text=f'{item_sum} ({"+" if item_sum>target else "-"}{abs(item_sum-target)})', font=('Arial', 18))
+        h *= (item_sum / target) if target != 0 else 0
+        self.canvas.create_rectangle(x, y, x + w, y + h, fill="black")
+        diff = abs(item_sum - target)
+        sign = "+" if item_sum > target else "-"
+        self.canvas.create_text(
+            x + w // 2,
+            y + h + screen_padding,
+            text=f"{item_sum} ({sign}{diff})",
+            font=("Arial", 18),
+        )
 
-    def draw_genome(self, genome, gen_num):
-        for i in range(num_items):
-            item = self.items_list[i]
-            active = genome[i]
-            item.draw(self.canvas, active)
+    def draw_result_text(self, sum_val):
         x = (self.width - screen_padding) / 8 * 6
         y = screen_padding
         w = (self.width - screen_padding) / 8 - screen_padding
         h = self.height / 4 * 3
-        self.canvas.create_text(x + w, y + h + screen_padding*2, text=f'Generation {gen_num}', font=('Arial', 18))
+        self.canvas.create_text(
+            x + w,
+            y + h + screen_padding * 2,
+            text=f"Result sum: {sum_val}",
+            font=("Arial", 18),
+        )
 
     def run(self):
-        global pop_size
-        global num_generations
+        # DP-based solver for a "closest subset sum" problem
+        values = [item.value for item in self.items_list]
+        target = self.target
+        max_sum = sum(values)
+        logger.debug(
+            f"Starting DP computation. Target={target}, Max sum of all items={max_sum}"
+        )
 
-        def gene_sum(genome):
-            total = 0
-            for i in range(len(genome) - 1):
-                if genome[i]:
-                    total += self.items_list[i].value
-            return total
+        dp = [[False] * (max_sum + 1) for _ in range(num_items + 1)]
+        dp[0][0] = True
 
-        def fitness(genome):
-            return abs(gene_sum(genome) - self.target)
+        # Fill the DP table
+        for i in range(1, num_items + 1):
+            val = values[i - 1]
+            for s in range(max_sum + 1):
+                if dp[i - 1][s]:
+                    dp[i][s] = True
+                    if s + val <= max_sum:
+                        dp[i][s + val] = True
 
-        def get_population(last_pop=None, fitnesses=None):
-            population = []
-            if last_pop is None:
-                for g in range(pop_size):
-                    genome = []
-                    for bit in range(num_items):
-                        genome.append(random.random() < frac_target)
-                    population.append(genome)
-                return population
-            else:
-                # elitism
-                elites = []
-                for e in range(elitism_count):
-                    elites.append(fitnesses[e])
-                for e in last_pop:
-                    if fitness(e) in elites:
-                        population.append(e)
+            logger.debug(f"After processing item {i} (value={val}), DP row completed.")
 
-                def select_parents(min_fitness):
-                    weights = []
-                    for parent in last_pop:
-                        if fitness(parent) == 0.0:
-                            weights.append(1.0)
-                        else:
-                            weights.append(min_fitness / fitness(parent))
+        # Find closest sum
+        closest_sum = None
+        for diff in range(max_sum + 1):
+            if target - diff >= 0 and dp[num_items][target - diff]:
+                closest_sum = target - diff
+                break
+            if target + diff <= max_sum and dp[num_items][target + diff]:
+                closest_sum = target + diff
+                break
 
-                    def get_by_weight():
-                        idx = random.randint(0, pop_size - 1)
-                        while random.random() < weights[idx]:
-                            idx = random.randint(0, pop_size - 1)
-                        return last_pop[idx]
+        logger.debug(f"Closest sum to target found: {closest_sum}")
 
-                    return get_by_weight(), get_by_weight()
+        # Backtrack to find chosen items
+        chosen_indices = []
+        s = closest_sum
+        for i in range(num_items, 0, -1):
+            val = values[i - 1]
+            if s - val >= 0 and dp[i - 1][s - val]:
+                chosen_indices.append(i - 1)
+                s -= val
 
-                def crossover(parent1, parent2):
-                    length = len(parent1)
-                    x = random.randint(0, length // 2)
-                    y = x + length // 2
-                    g_out = []
-                    for i in range(length):
-                        if x < i <= y:
-                            g_out.append(parent2[i])
-                        else:
-                            g_out.append(parent1[i])
-                    if len(g_out) < num_items:
-                        print('Error!')
-                    return g_out
+        chosen_indices.reverse()  # for readability
+        chosen_values = [values[i] for i in chosen_indices]
+        logger.debug(f"Chosen subset indices: {chosen_indices}")
+        logger.debug(f"Chosen subset values: {chosen_values}")
+        logger.debug(f"Chosen subset sum: {sum(chosen_values)}")
 
-                def mutate(g_in):
-                    x = random.randint(0, len(g_in) - 1)
-                    g_out = []
-                    for i in range(len(g_in)):
-                        if i == x:
-                            g_out.append(not g_in[i])
-                        else:
-                            g_out.append(g_in[i])
-                    return g_out
-
-                # fill generation with new individuals
-                while len(population) < pop_size:
-                    # select two random parents by weighted selection
-                    # note no guarantee of uniqueness - could get the same parent twice
-                    parents = select_parents(fitnesses[0])
-                    # perform crossover to generate new individual
-                    baby = crossover(parents[0], parents[1])
-                    # potentially perform mutation
-                    if random.random() < mutation_rate:
-                        baby = mutate(baby)
-                    # add to next generation
-                    population.append(baby)
-
-                return population
-
-        def generation_step(generation=0, pop=None):
-            if generation >= num_generations:
-                return  # Stop the process after the set number of generations
-
-            if pop is None:
-                pop = get_population()
-
-            fitnesses = []
-            best_of_gen = None
-            min_fitness = 9999
-            for genome in pop:
-                fit = fitness(genome)
-                if fit < min_fitness:
-                    best_of_gen = genome
-                    min_fitness = fit
-                fitnesses.append(fit)
-            fitnesses.sort()
-
-            print(f'Best fitness of generation {generation}: {min_fitness}')
-            print(best_of_gen)
-            print()
-
-            # Schedule the UI updates in the main thread
-            self.after(0, self.clear_canvas)
-            self.after(0, self.draw_target)
-            self.after(0, self.draw_sum, gene_sum(best_of_gen), self.target)
-            self.after(0, self.draw_genome, best_of_gen, generation)
-
-            # Schedule the next generation step after a delay, unless we're at the global optimum (fitness == 0)
-            if fitnesses[0] != 0:
-                self.after(int(sleep_time * 1000), generation_step, generation + 1, get_population(pop, fitnesses))
-
-        # Start the evolutionary process
-        generation_step()
+        self.after(0, self.clear_canvas)
+        self.after(0, self.draw_target)
+        self.after(0, self.draw_sum, closest_sum, target)
+        self.after(0, self.draw_items, chosen_indices)
+        self.after(0, self.draw_result_text, closest_sum)
 
 
-# In python, we have this odd construct to catch the main thread and instantiate our Window class
-if __name__ == '__main__':
-    UI()
+if __name__ == "__main__":
+    app = UI()
+    app.mainloop()
